commands: add a --samba-debug-level opt & var for debugging
Add --samba-debug-level cli option, and a matching SAMBA_DEBUG_LEVEL env var so
that the samba processes we start from sambacc can be debugged. This is done on
a per-command basis (per- samba-container command) so that multiple containers
consuming the same config can be debugged independently. You can still add
debug lines to the smb.conf that gets generated.



Signed-off-by: John Mulligan <jmulligan@redhat.com>