commands: add --skip-if-file option for simple conditional ops
In order to keep operations low on scripting at the container orch
layer, we add an option to skip the rest of the command if a
sentinel file (any file really) exists.

This will allow us to continue using `samba-container` as a building
block for batches of commands when we need certain actions to be
skipped if another container has already done something.


Signed-off-by: John Mulligan <jmulligan@redhat.com>