commands: when ctdb enabled, register dns only if leader
When the container is enabled for ctdb we will only perform
the dns registration step if the current "node" is the cluster leader.
What that really means is deliberately obscured by the leader locator.
In practice it should mean that only 1 node of N in a clustered
group of containers will be responsible for running the
'net ads dns register ...' commands.


Signed-off-by: John Mulligan <jmulligan@redhat.com>