container: allow building for external consumption
Previously, we built the (sdist, wheel) packages in the source tree.
Getting the files out of the container was a matter of mounting
in a dir for the source tree. Now, we support having a "named distribution"
that is explicitly for consumption by additional containers or by
mounting in just "/srv" or "/srv/dist" to get the build artifacts.


Signed-off-by: John Mulligan <jmulligan@redhat.com>