pyproject: save version to python file
This file is not currently used but could be used in the future
so that the package "knows" its own version (and could be displayed
in logs, etc). I was not able to store the full hash but the short
hash stored in the version string should be good enough in a pinch.


Signed-off-by: John Mulligan <jmulligan@redhat.com>