sambacc: small fix to output of get_samba_specifics
Previously, if the env var SAMBA_SPECIFICS was unset or empty
the function would return a set with one item, the empty string ("").
When I went to add a unit test for the function I assumed it would
return an empty set but it did not. This change makes the function
behave as I would expect it to when SAMBA_SPECIFICS is unset.


Signed-off-by: John Mulligan <jmulligan@redhat.com>