pyproject.toml: downgrade required setuptools_scm version
In order to build an RPM directly from a source tarball on fedora 35 it
appears that we can't use the current version of setuptools_scm.
After some testing-by-hand I don't see any immediate downsides to using
the older version. It's a little hacky to go backwards but now's the
best time to do so since we have never released yet.


Signed-off-by: John Mulligan <jmulligan@redhat.com>