tox.ini: reorganize tox config for more granular env selection
Break the tox configuration up so that the formatting, mypy, and
pytest runs occur in their own environments. Futher, allow py39
based envs to operate with fewer dependencies so that we still
ensure we work on py39 but the "fully featured" test run
occurs on whatever the "native" py3 version is.


Signed-off-by: John Mulligan <jmulligan@redhat.com>