sambacc: make permissions module core work without xattr module
The xattr module works well if testing with the python3 packages from the
system but easily breaks when sing tox to test other python versions. This can
be resolved either by letting python compile the C code needed for xattr or by
skipping xattr support if it is not the available on our current python
version.  This patch implements the latter.


Signed-off-by: John Mulligan <jmulligan@redhat.com>