setup: make sure to include sambacc.commands in packages list
Strangely, the packaging was working when running the build from an active
repo (git, when setuptools_scm enabled) but sambacc.commands was getting
excluded from builds from source tarballs. So it seems that setuptools_scm
was saving us when run from repo but not when run from the tarball.
This small change to the setup.cfg appears to be enough to fix the wheel
file that gets generated from the tarball.


Signed-off-by: John Mulligan <jmulligan@redhat.com>