sambacc: complete type annotations for container_dns.py
This includes a small change to the signature of the watch function
to only accept a print_func callable and never none. This required
a small update to the test code. The real code always passed a function
and the ability to pass none never really could have worked right
anyway.


Signed-off-by: John Mulligan <jmulligan@redhat.com>