pyproject: require type annotations for sambacc function defs
The sambacc dir is basically a library and we want to ensure high
quality and good documentation so configure mypy to require
type annotations on functions defined in sambacc.
The sambacc.commands on the other hand are mean to be only
run as part of our CLI tools so we'll be a bit less strict there.


Signed-off-by: John Mulligan <jmulligan@redhat.com>