container: update build.sh with ability to install dependencies
Most of the logic around building and testing sambacc was in the
build.sh script already, so why not include the ability to install
the os packages too.

When called with `--install` as the first argument instead of a revision
id, we assume that the script is being called specifically and only to
install the package dependencies.

Signed-off-by: John Mulligan <jmulligan@redhat.com>