schema: add new json-schema based schema files to the tree
Add two schema files, one in yaml and one in json, to the tree.
The YAML file is the file to be edited, but in cases where yaml
is an extra dependency we also provide a conversion to raw json.
These yaml files help describe the configuration and may be later
used with validation tools or documentation generators.

Signed-off-by: John Mulligan <jmulligan@redhat.com>