commands: add cli option & env var to control config validation
Add a --validate-config cli option and SAMBACC_VALIDATE_CONFIG
environment variable to enable/disable (or leave automatic) the json
schmea based validation of the config. Because json schema is new and
there are cases (primarily as a dev hacking on stuff) we give the
ability to turn off the validation. Note - the validation is auto by
default - if the library is not present the validation will be
automatically skipped.

Signed-off-by: John Mulligan <jmulligan@redhat.com>