tox: add schemacheck to verify that generated schema data matches yaml
The default behavior of tool.py in the schema dir is to check that the
data in the yaml based files match that in the JSON & py files.
Add a tox rule to invoke the tool to help ensure changes to yaml file
aren't being made without corresponding updates to the others.

Signed-off-by: John Mulligan <jmulligan@redhat.com>