container: always run rpm build as part of a normal build
Previously, we only executed the rpm build if the `distname` was given,
meaning we expected to output the rpms to a "distribution" location.
But this meant that the rpms were typically only built when samba
containers are being built and never as part of a CI run.
This change updates the build script to always build the RPMs, even
if distname is not supplied by using a "scratch" dist dir.
Now, the rpm build should be executed as part of a typical CI run.

Signed-off-by: John Mulligan <jmulligan@redhat.com>