sambacc: ignore mypy errors if it can't import jsonschema
Due to a somewhat hilarious misalignment of the stars the recently added
jsonschema support doesn't actually work on the distro we build for.
This was masked by the fact that the sambacc CI didn't test the distro
level stuff that actually showed there was a problem.
This change just tells mypy to chillax if it fails to import typing data
for jsonschema so that it can pass the tests with or without the
jsonschema library present (as it is an optional dep).

Signed-off-by: John Mulligan <jmulligan@redhat.com>