container: inform git that our repo path is acceptable
In some cases, specifically the github CI runner, calls made to
git by setuptools_scm were failing leading to source taballs that
were missing files, which lead to rpmbuild failures. This was all
because of "stricter repository ownership checks" in git.
Use the `safe.directory` configuration key to always permit
the use of our checkout when run in the container.

Signed-off-by: John Mulligan <jmulligan@redhat.com>