sambacc: make config module robust to unexpected jsonschema versions
If jsonschema module doesn't understand our schemas then we assume that
the version is too old and we should treat it the same as not having
jsonschema installed. We recently found that having older jsonschema
versions present in the container actually breaks sambacc. Because we
prefer to install from rpm packages rather than pip in samba-container
we ended up with older, and an untested version, of the library. This
change tries to make the sambacc code more robust in that case.

Signed-off-by: John Mulligan <jmulligan@redhat.com>