container: generate a more robust rpm spec file
When using the container to build "binary" rpms passing the pversion and
rversion vars mostly worked, but the srpm it generated was not usable.
This change generates a new spec file based on the old, but with
pversion and rversion predefined, based on the same setuptools_scm
values, such that the srpm is usable in other environments.

Signed-off-by: John Mulligan <jmulligan@redhat.com>