sambacc: do not raise a KeyError if a config lacks globals
In the case of AD DC it's pretty normal for there to be no globals key
in the instance configuration. If we want to optionally support global
configuration params for addc it's convenient to simply return an empty
list rather than raise an exception in this case.  If the globals list
refers to global configs that don't exist exceptions will continue to be
raised which seem like the right thing to do.

Signed-off-by: John Mulligan <jmulligan@redhat.com>