tests: add unit tests for smbconf_samba
Add unit(ish) tests for smbconf_samba. We have to work around an (IMO)
annoying behavior of samba's. When using the registry backend once the
registry tdb file is opened it is never closed and subsequent calls
to smbconf's init_reg or similar end up reusing whatever first tdb file
got opened. (I looked at samba master code, this is still the case
AFAICT.) So we do a bit of hackery with regards to test fixtures to
get a "clean" SMBConf for each test. This shouldn't be much of an
issue outside of tests as you typically only have one smb.conf and/or
one data dir with a registry.tdb in it.

Signed-off-by: John Mulligan <jmulligan@redhat.com>