commands: enable url and rados openers for sambacc config
When building the initial sambacc configuration, make sure
we support passing URL and optionally enable rados support.

To support passing urls/uris no changes to the CLI parsing needs
to happen but the env var parsing previously assumed a colon-separated
list of paths much like the PATH env var. Keep that as a viable
option for backwards compatibility but also allow the env var to
contain a JSON-formatted list of strings. Example:

```
[0] $ export SAMBACC_CONFIG='["http://localhost:8080/my/config.json", "/my/config.json"]'
[0] $ samba-container --identity demo <...>
```

Signed-off-by: John Mulligan <jmulligan@redhat.com>