sambacc: fix join interactive test
Fix how the Joiner and the test for interactive login was using stdin.
When the test was run without pytest capturing the stdio, the test would
hang waiting on input. This behavior was incorrect. Add a hook to the
Joiner for the unit tests to use so that the stdin of the interactive
join command can be manipulated.

Signed-off-by: John Mulligan <jmulligan@redhat.com>