sambacc: extend rados pseudo-uris to include getting mon config keys
There are cases where certain configuration items may not be stored in
the Ceph cluster's  rados pool but rather in the Ceph MON's
configuration key store. This can be accessed via a mon command. Add
support for new pseudo uri's formed like:
  rados:mon-config-key:foo/bar/baz

Instead of fetching an item from a rados pool, this will try to fetch a
key in a manner equivalent to running the following CLI command:
  `ceph config-key get foo/bar/baz`

This idea is borrowed from the ceph iscsi support.

Signed-off-by: John Mulligan <jmulligan@redhat.com>