sambacc: add ClusterMeta protocol to ctdb functions
Add ClusterMeta & ClusterMetaObject classes to begin abstracting some of
the code paths that were previously always assuming a file containing
JSON on disk that could be locked. This change should be largely
invisible to the commands using ctdb.py. In the future, we will use the
more abstracted versions so that sambacc will be able to be more
flexible WRT how the metadata is stored as instructed by an
orchestration system.

Signed-off-by: John Mulligan <jmulligan@redhat.com>