tox.ini: update version of black formatter tool
Allow tox to use a version of the black tool from 2024.  Having added
doc strings to protocols the black tool, using the 2024 style,  will no
longer rewrite long-ish protocol method definitions that will span
multiple lines either way from:

    def foo(a: str, b: str, c: SomethingCompletelyDifferent) -> int:
        ...

to

    def foo(
        a: str, b: str, c: SomethingCompletelyDifferent
    ) -> int: ...

Just to force the ellipsis to be on the "same line" as the definition.
This is change I find a bit unnecessary and a tad bit ugly.  I did
report this to the black team but they chose to make the change anyway.

So they've unintentionally caused me to write more doc strings. Which is
a good thing I guess. :-)

Signed-off-by: John Mulligan <jmulligan@redhat.com>