sambacc: replace path key with uri key for ctdb metadata state object
The sambacc system has a "meta-management" JSON file that it uses for
interacting with CTDB. Currently, this must be a file in the file
system. However, future changes are going to make supporting different
back-ends and protocols possible. Rename the key in the config to
reflect that fact.

NB: This state file gets written to by sambacc and thus not all the
same protocol for config files and such are supported here.

Signed-off-by: John Mulligan <jmulligan@redhat.com>