commands: prepare ctdb commands for using cluster meta objects
This change prepares the ctdb related commands in sambacc for using
different paths or, eventually, different uris for the cluster metadata
state data. Currently, this must still be a file but the file can be
specified by path `/foo/bar/baz` or uri style `file:///foo/bar/baz`.

The CLI now constructs a ClusterMeta object and passes it to the library
functions rather than passing a path.

Signed-off-by: John Mulligan <jmulligan@redhat.com>