sambacc: add get_object method to rados handler class
Add a method for getting a RADOSObjectRef based on a uri-like string,
similar to the `rados_open` method that exists to support the urllib
functionality. This includes a small refactor of the uri/url parsing
so that it can be shared for both `get_object` and `rados_open`.
Note that get_object advertises returning RADOSObjectRef so that
special specific-to-rados APIs can be used easily.

Signed-off-by: John Mulligan <jmulligan@redhat.com>