commands: add --archive option to ctdb-migrate cmd
Passing this argument with a path will archive any of the tdb files that
get migrated to ctdb to the specified path. This moves files from the
typical location to aid in debugging (it's more obvious that this
instance is using ctdb) and for comparison in case something needs to
be compared post-conversion.

Signed-off-by: John Mulligan <jmulligan@redhat.com>