sambacc: rename function to manage_cluster_meta_updates
Rename a function to manage_cluster_meta_updates to better match what it
does. A future change will want to use the word "monitor" for something
else and it would be confusing to continue calling this a "monitor"
function vs "manage -- and this function is the more active
"participant" of the two.

Signed-off-by: John Mulligan <jmulligan@redhat.com>