sambacc: add monitor_cluster_meta_changes function
Add the monitor_cluster_meta_changes to the ctdb module. This function
monitors the cluster meta and can reflect those changes into the ctdb
nodes file and/or call reload nodes. Unlike the previous
manage_cluster_meta_updates the sambacc process is not an active
participant in the management of the cluster meta rather we assume some
outside entity is updating the metadata with the desired state of the
cluster correctly. monitor_cluster_meta_changes exists to translate
that state into something ctdb can use.

Signed-off-by: John Mulligan <jmulligan@redhat.com>