commands: add ctdb-monitor-nodes command
Add a `ctdb-monitor-nodes` command - this command will monitor the
cluster meta and update the containers/ctdb when/if it changes.
The `--write-nodes` parameter is used to instruct the command to
write the ctdb nodes file on changes. The `--reload` option is used to
control what "nodes" issue `ctdb reloadnodes` commands: leader (only the
cluster leader), all (all nodes that see the cluster meta change), and never
(never issue reloadnodes commands).

Signed-off-by: John Mulligan <jmulligan@redhat.com>