commands: fix a mypy failure in the ctdb commands module
It seems that a recent mypy release is now determining that certain
family types (not AF_INET, or AF_INET6) may return an address tupe
with the first element type being an int rather than a str. The
(expanded) revealed type is:
```
tuple[
    socket.AddressFamily,
    socket.SocketKind,
    builtins.int,
    builtins.str,
    Union[
        tuple[builtins.str, builtins.int],
        tuple[builtins.str, builtins.int, builtins.int, builtins.int],
        tuple[builtins.int, builtins.bytes]
    ]
]
```

This change ensures that we only check the family values we care about
and use a type assertion for those families where we expect the address
value to be a str.

FWIW I could not find any documentation for what families have an
integer in the first element of the addr tuple. Looking at the python
sources it seems like it could be AF_NETLINK, AF_QIPCRTR, AF_VSOCK -
all families I have never seen in the wild before :-)

Signed-off-by: John Mulligan <jmulligan@redhat.com>