commands: invalidate cached config when node is not ctdb leader
When the `samba-container config-update --watch` command is run
on a CTDB enabled server we avoid updating the local config unless
the node is the ctdb leader. However, the update function was returning
the current config which the watch loop assigned to the previous
var, meaning that the current and previous config would be equal on
future trips around the loop. Therefore once the node became the leader
it would not update the config even if it differed from the config
in samba. This is a problem on start up because the loop initially
runs before ctdb is ready and the node leadership is not known.
Change it so that when the node is not a leader it returns None
and "unsets" the previous config. This could lead to extra unneeded
updates but should avoid missing updates that are needed.

Signed-off-by: John Mulligan <jmulligan@redhat.com>