tox.ini: add new environments for generating grpc related files
Add a new tox environment that is to be used for regenerated python
source files from .proto files. Currently, it is up to the contributor
to make sure these are up-to-date.

A second env is provided for use on centos/fedora with packaged versions
for grpc and protobuf in order to make sure the generated files match
what the planned deployment containers will support.

Signed-off-by: John Mulligan <jmulligan@redhat.com>