varlink: add a varlink interface file for a keybridge service
Define a simple-ish interface for fetching data blobs from a peer
service. The intent is that the backing service can be a basic
key-value store or a more complex secret store. The keybridge
server encapsulates the complexity of talking to a web-service or
whatnot requiring the client to only deal with identifying the
scope and key to fetch the value.
Plus debugging and introspection stuff.

Signed-off-by: John Mulligan <jmulligan@redhat.com>