workflows: Dynamically detect fedora versions for test matrix
Instead of updating the versions during each fedora release we could
automatically filter out the recent versions from publically available
json data for fedora releases. In addition to that we rename fedora
test jobs to symbolically represent the supported releases of fedora.
This is done in an attempt to reduce the need for updating mergify
configuration during each fedora release.

Signed-off-by: Anoop C S <anoopcs@samba.org>