commands/satellite: add commands for satellite servers & keybridge
Instead of constantly having to define new dedicated subcommands for
things that don't fit neatly under the `samba-container` umbrella,
create a new concept for things provided by sambacc that "orbit" the
samba & smbd processes but are very optional in most scenarios.

Establish the keybridge server as our first satellite. The keybridge
server can be set up via the sambacc config or configured/customized
by command line values. In particular, the TLS credentials needed for
KMIP can be specified on the command line while the other params
are in the file because of ceph-shaped reasons.

Signed-off-by: John Mulligan <jmulligan@redhat.com>