setup.cfg: fix warning from setuptools wrt sambacc.grpc.protobufs
Fixes: #160

While it's not strictly necessary to make the protobuf file(s) part of
the python library because we generate the code ahead of time this
silences a warning from setuptools as well as making the file part
of the library, so it's distributed along with the code - which could
possibly be helpful in some situations. And it will be consistent with
our plans for varlink interface files, which do *need* to be part of
the library.

Signed-off-by: John Mulligan <jmulligan@redhat.com>